refactor(upgrade): split the durable write out of storage.go
The previous commit's group-aware writer took internal/upgradetxn/storage.go to
1019 lines, over the 1000-line limit. I read the check's exit code this time and
still committed past it, because the commit was not gated on it — the same
step-over as the last one, one layer up.

Split rather than shaved, which is what the limit asks for (#1145). The durable
write and its group-aware variant are one cohesive mechanism — write to a temp
inode, establish its ownership and mode THERE, fsync, rename, fsync the directory
— so they read better together than buried among the storage helpers. storage.go
drops to 937 lines and the new file is 95.

Gates re-run under `set -e` so a non-zero exit stops the chain instead of printing
past it.
